#564 :recycle:, :truck: refactor file structure
