+++ wls-gui-wahllokalsystem/src/api/wls-clients/broadcast-service/BroadcastMessage.ts
